balance is a decimal

+++ db/migrate/20161124190638_add_balance_to_account.rb
@@ -1,5 +1,5 @@
 class AddBalanceToAccount < ActiveRecord::Migration[5.0]
   def change
-    add_column :accounts, :balance, :integer, default: 0, null: false
+    add_column :accounts, :balance, :decimal, default: 0, null: false
   end
 end
